+++ CHANGELOG.md
@@ -6,6 +6,8 @@ empty string.
 * Deprecated all `Angel` methods that directly touch an `HttpServer` or `HttpRequest`.
 * Created the `AngelHttp` class.
 * Removed explicit dependence on `dart:io` for `Angel`, `RequestContext`, `ResponseContext`.
+* Created `lib/http.dart`, which exports HTTP-specific functionality.
+* `AnonymousService` now takes `FutureOr`.
 
 # 1.1.0+3
 * Modified `ResponseContext#isOpen` to also return `false` if streaming is being used.